switch DataUsageInfo to interface
